Adds binary search parameters to basic search and MCMC search
Note: this also adds a check for transient, if not called, then the
whole data is searched. This does not however mean only transient, but
also glitch like searches which use the transient tools.

Additionally fixes the plot_walkers command to allow it to plot walkers
when ndim =1 and to switch off the use of offsets