Forces output of Fstat to be finite and non-nan
It seems that , due to the choice of transient windows, occasionally
the calculated Fstat is +/- inf or a nan. This behaviour is not
understood. To address this, we apply a simple check and if it is an
undesirable value, return 0. This will inherently bias the chains away
from regions of parameter space in which the output breaks.