Improves pruning method
If an initialisation step is callled, the sampler runs the simulations
for nsteps and selects the point with the highest lnprobability from
which to restart the simulations. Previously, only the last step was
used. In this update all steps are used and a message is printed to
allow the user to determine if the selected point is appropriate.