Renames sftfilepath -> sftfilepattern
This renames the input sftfilepath to sftfilepattern and adds
documentation on how that should be used, i.e. a colon separated list of
wildstring or exact matches. In globbing for all matches, the colon
split is added in. sftfilepath is still used by `Writer` since
an exact path is known.