Adds parallelisation of the artifact SFT generation
Since each SFT is generated in a for loop, the process can be quite
slow. This allows the user with `pathos` multiprocessing installed to
leverage `-N` cores (set by command line arg) to speed the process up
Setting N=1 or if pathos is not installed, reverts to the usual
behaviour.